move monotonic checks up
